requirements.txt
numpy==1.24.3
pytest==7.3.1
urllib3==1.26.7
allure-pytest==2.13.2
pymysql==1.0.3
python-dotenv==0.21.1
webdriver-manager==3.8.6
selenium==4.9.0
pandas==2.0.2
openpyxl==3.1.2
pytest-xdist==3.3.1
pytest-rerunfailures==11.1.2



# allure-python-commons==2.13.2
# mysql-connector-python==8.0.33
# pytest-base-url==2.0.0
# python-dateutil==2.8.1
# #python-dateutil==2.8.2
# python-dotenv==0.21.1
# #python-dotenv==1.0.0
# python-slugify==6.1.2
# requests==2.28.1
# selenium==4.3.0
# webdriver-manager==3.8.6
# websockets==10.1
# pymysql==1.0.3
# exceptiongroup==1.1.1 #Add by installing xdist
# execnet==1.9.0        #Add by installing xdist
# iniconfig==2.0.0      #Add by installing xdist 
# pluggy==1.0.0         #Add by installing xdist
# psutil==5.9.5         #Add by installing xdist
# pytest-xdist==3.3.1   #Add by installing xdist
# tomli==2.0.1          #Add by installing xdist
# pytz==2023.3
# tzdata==2023.3
# openpyxl==3.1.2
# et-xmlfile==1.1.0
# pytest-rerunfailures==11.1.2